Add YuYu multsrc
