+++ src/pt/egotoons/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Ego Toons'
+    extClass = '.EgoToons'
+    themePkg = 'yuyu'
+    baseUrl = 'https://egotoons.com'
+    overrideVersionCode = 0
+    isNsfw = true
+}
+
+apply from: "$rootDir/common.gradle"
