+++ src/pt/galinhasamuraiscan/src/eu/kanade/tachiyomi/extension/pt/galinhasamuraiscan/GalinhaSamuraiScan.kt
@@ -1,15 +1,17 @@
 package eu.kanade.tachiyomi.extension.pt.galinhasamuraiscan
 
-import eu.kanade.tachiyomi.multisrc.madara.Madara
-import java.text.SimpleDateFormat
-import java.util.Locale
+import eu.kanade.tachiyomi.multisrc.yuyu.YuYu
+import eu.kanade.tachiyomi.network.interceptor.rateLimit
 
-class GalinhaSamuraiScan : Madara(
+class GalinhaSamuraiScan : YuYu(
     "Galinha Samurai Scan",
     "https://galinhasamurai.com",
     "pt-BR",
-    dateFormat = SimpleDateFormat("dd/MM/yyyy", Locale("pt", "BR")),
 ) {
-    override val useLoadMoreRequest = LoadMoreStrategy.Always
-    override val useNewChapterEndpoint = true
+    override val client = super.client.newBuilder()
+        .rateLimit(2)
+        .build()
+
+    // Moved from Madara to YuYu
+    override val versionId = 2
 }
