+++ src/pt/nekotoons/build.gradle
@@ -1,0 +1,10 @@
+ext {
+    extName = 'Neko Toons'
+    extClass = '.NekoToons'
+    themePkg = 'yuyu'
+    baseUrl = 'https://nekotoons.site'
+    overrideVersionCode = 0
+    isNsfw = true
+}
+
+apply from: "$rootDir/common.gradle"
