+++ src/pt/plumacomics/build.gradle
@@ -1,9 +1,9 @@
 ext {
     extName = 'Pluma Comics'
     extClass = '.PlumaComics'
-    themePkg = 'madara'
-    baseUrl = 'https://plumacomics.cloud'
-    overrideVersionCode = 0
+    themePkg = 'yuyu'
+    baseUrl = 'https://new.plumacomics.cloud'
+    overrideVersionCode = 41
     isNsfw = false
 }
