+++ src/pt/spectralscan/src/eu/kanade/tachiyomi/extension/pt/spectralscan/SpectralScan.kt
@@ -1,16 +1,17 @@
 package eu.kanade.tachiyomi.extension.pt.spectralscan
 
-import eu.kanade.tachiyomi.multisrc.madara.Madara
-import java.text.SimpleDateFormat
-import java.util.Locale
+import eu.kanade.tachiyomi.multisrc.yuyu.YuYu
+import eu.kanade.tachiyomi.network.interceptor.rateLimit
 
-class SpectralScan : Madara(
+class SpectralScan : YuYu(
     "Spectral Scan",
     "https://spectralscan.xyz",
     "pt-BR",
-    dateFormat = SimpleDateFormat("MMMM d, yyyy", Locale("pt", "BR")),
 ) {
-    override val useLoadMoreRequest = LoadMoreStrategy.Never
-    override val useNewChapterEndpoint = true
-    override val mangaDetailsSelectorStatus = "div.post-content_item:contains(Estado) > div.summary-content"
+    override val client = super.client.newBuilder()
+        .rateLimit(2)
+        .build()
+
+    // Moved from Madara to YuYu
+    override val versionId = 2
 }
